Reorganize the tests
Created home.tests folder, and split all the tests as follows:
	deleted:    home/tests.py
	new file:   home/tests/test_display_question_feature.py
	new file:   home/tests/test_explore_page.py
	new file:   home/tests/test_insert_question_feature.py
	new file:   home/tests/test_profile_middleware.py
	new file:   home/tests/test_question_tag_model_functions.py
	new file:   home/tests/test_tags_page.py
	new file:   home/tests/test_login.py
