+++ home/tests/test_profile_middleware.py
@@ -1,0 +1,30 @@
+import pytest
+from home.profile_middleware import ProfileMiddleware
+from unittest.mock import Mock
+from django.contrib.auth.models import User
+from home.models import Profile
+
+
+class TestProfileMiddleware():
+    class LoggedOutUserMock:
+        is_authenticated = False
+
+    @pytest.mark.django_db
+    def test_process_request_with_logged_in_user(self):
+        logged_in_user = User.objects.latest('id')
+        pm = ProfileMiddleware('response')
+        assert logged_in_user.is_authenticated is True
+
+        request_mock = Mock(user=User.objects.latest('id'))
+        pm.process_view(request_mock, None, None, None)
+        assert request_mock.profile == Profile.objects.filter(user=logged_in_user).first()
+
+    @pytest.mark.django_db
+    def test_process_request_with_logged_out_user(self):
+        logged_out_user = self.LoggedOutUserMock()
+        pm = ProfileMiddleware('response')
+        assert logged_out_user.is_authenticated is False
+
+        request_mock = Mock(user=self.LoggedOutUserMock())
+        pm.process_view(request_mock, None, None, None)
+        assert request_mock.profile is None
